Make Route for each season
