+++ src/component/index.js
@@ -1,10 +1,10 @@
 import Header from "./Header"
-import Player from "./Player"
+import Players from "./Players"
 import Developer from "./Developer"
 import LastScore from "./LastScore"
 import SideBar from "./SideBar"
 import SideBarFake from "./SideBarFake"
 
 export {
-    Header,Player,Developer,LastScore,SideBar,SideBarFake
+    Header,Players,Developer,LastScore,SideBar,SideBarFake
 }
